feat: meta procedure options
