+++ src/common/procedure/Cargo.toml
@@ -7,15 +7,16 @@ license.workspace = true
 [dependencies]
 async-trait.workspace = true
 async-stream.workspace = true
+backon = "0.4.0"
 common-error = { path = "../error" }
 common-runtime = { path = "../runtime" }
 common-telemetry = { path = "../telemetry" }
 futures.workspace = true
+humantime-serde = "1.1"
 object-store = { path = "../../object-store" }
 serde.workspace = true
 serde_json = "1.0"
 smallvec = "1"
-backon = "0.4.0"
 snafu.workspace = true
 tokio.workspace = true
 uuid.workspace = true
